requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.5
decorator==5.1.1
Django==4.1.6
django-anymail==9.0
django-appconf==1.0.5
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-ipware==4.0.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==1.1.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.31
pure-eval==0.2.2
Pygments==2.13.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
traitlets==5.4.0
tzdata==2022.2
urllib3==1.26.14
wcwidth==0.2.5
